Split patap into different modules.
